requirements.txt
Pillow==9.4.0
requests==2.28.1
streamlit==1.18.1
streamlit_lottie==0.0.3
